Move DisplayLinkObserver to Utilities
